// give-away/src/api/models/CartRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Give Away API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PaymentMethod } from './PaymentMethod';
import {
    PaymentMethodFromJSON,
    PaymentMethodFromJSONTyped,
    PaymentMethodToJSON,
} from './PaymentMethod';

/**
 * 
 * @export
 * @interface CartRequest
 */
export interface CartRequest {
    /**
     * 
     * @type {PaymentMethod}
     * @memberof CartRequest
     */
    paymentMethod?: PaymentMethod;
    /**
     * 
     * @type {string}
     * @memberof CartRequest
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CartRequest
     */
    recipientName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CartRequest
     */
    phone?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof CartRequest
     */
    listItemId?: Array<string> | null;
}

/**
 * Check if a given object implements the CartRequest interface.
 */
export function instanceOfCartRequest(value: object): value is CartRequest {
    return true;
}

export function CartRequestFromJSON(json: any): CartRequest {
    return CartRequestFromJSONTyped(json, false);
}

export function CartRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CartRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'paymentMethod': json['paymentMethod'] == null ? undefined : PaymentMethodFromJSON(json['paymentMethod']),
        'address': json['address'] == null ? undefined : json['address'],
        'recipientName': json['recipientName'] == null ? undefined : json['recipientName'],
        'phone': json['phone'] == null ? undefined : json['phone'],
        'listItemId': json['listItemId'] == null ? undefined : json['listItemId'],
    };
}

export function CartRequestToJSON(value?: CartRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'paymentMethod': PaymentMethodToJSON(value['paymentMethod']),
        'address': value['address'],
        'recipientName': value['recipientName'],
        'phone': value['phone'],
        'listItemId': value['listItemId'],
    };
}